+++ apps/web/src/entities/user/index.ts
@@ -11,7 +11,7 @@ export {
   mypageQueryKeys,
   useAnalyticsQuery,
   useMypageQuery,
-} from './model/queries';
+} from './model/query-keys';
 export type {
   AnalyticsData,
   AnalyticsFocusEnvironment,
